Refactored sections to reduce LOC in main files
