+++ api/Products/mockup-data.js
@@ -1,0 +1,1529 @@
+module.exports = {
+  listMatchingProducts: {
+    ListMatchingProductsResult: {
+      Products: {
+        Product: {
+          Identifiers: {
+            MarketplaceASIN: {
+              MarketplaceId: 'ATVPDKIKX0DER',
+              ASIN: '059035342X'
+            }
+          },
+          AttributeSets: {
+            ItemAttributes: {
+              lang: 'en-US',
+              Author: 'Rowling, J.K.',
+              Binding: 'Paperback',
+              Brand: 'Scholastic Press',
+              Creator: {
+                Role: 'Illustrator',
+                Value: 'GrandPrÃ©, Mary'
+              },
+              Edition: '1st',
+              Feature: 'Recommended Age: 9 years and up',
+              ItemDimensions: {
+                Height: {
+                  Units: 'inches',
+                  Value: '0.80'
+                },
+                Length: {
+                  Units: 'inches',
+                  Value: '7.50'
+                },
+                Width: {
+                  Units: 'inches',
+                  Value: '5.20'
+                },
+                Weight: {
+                  Units: 'pounds',
+                  Value: '0.50'
+                }
+              },
+              IsAutographed: false,
+              IsMemorabilia: false,
+              Label: 'Scholastic Paperbacks',
+              Languages: {
+                Language: [
+                  {
+                    Name: 'english',
+                    Type: 'Unknown'
+                  },
+                  {
+                    Name: 'english',
+                    Type: 'Original Language'
+                  },
+                  {
+                    Name: 'english',
+                    Type: 'Published'
+                  }
+                ]
+              },
+              ListPrice: {
+                Amount: '10.99',
+                CurrencyCode: 'USD'
+              },
+              Manufacturer: 'Scholastic Paperbacks',
+              NumberOfItems: '1',
+              NumberOfPages: '320',
+              PackageDimensions: {
+                Height: {
+                  Units: 'inches',
+                  Value: '1.00'
+                },
+                Length: {
+                  Units: 'inches',
+                  Value: '7.50'
+                },
+                Width: {
+                  Units: 'inches',
+                  Value: '5.20'
+                },
+                Weight: {
+                  Units: 'pounds',
+                  Value: '0.50'
+                }
+              },
+              PackageQuantity: '1',
+              PartNumber: '9780590353427',
+              ProductGroup: 'Book',
+              ProductTypeName: 'ABIS_BOOK',
+              PublicationDate: '1999-10-01',
+              Publisher: 'Scholastic Paperbacks',
+              ReleaseDate: '1999-09-08',
+              SmallImage: {
+                URL: 'http://ecx.images-amazon.com/images/I/51MU5VilKpL._SL75_.jpg',
+                Height: {
+                  Units: 'pixels',
+                  Value: '75'
+                },
+                Width: {
+                  Units: 'pixels',
+                  Value: '51'
+                }
+              },
+              Studio: 'Scholastic Paperbacks',
+              Title: 'Harry Potter and the Sorcerer\'s Stone (Book 1)'
+            }
+          },
+          Relationships: null,
+          SalesRankings: {
+            SalesRank: [
+              {
+                ProductCategoryId: 'book_display_on_website',
+                Rank: '401'
+              },
+              {
+                ProductCategoryId: '15356791',
+                Rank: '5'
+              },
+              {
+                ProductCategoryId: '3153',
+                Rank: '8'
+              },
+              {
+                ProductCategoryId: '17468',
+                Rank: '16'
+              }
+            ]
+          }
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: '3b805a12-689a-4367-ba86-EXAMPLE91c0b'
+    }
+  },
+  getMatchingProduct: {
+    GetMatchingProductResult: {
+      ASIN: 'B002KT3XRQ',
+      status: 'Success',
+      Product: {
+        Identifiers: {
+          MarketplaceASIN: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            ASIN: 'B002KT3XRQ'
+          }
+        },
+        AttributeSets: {
+          ItemAttributes: {
+            lang: 'en-US',
+            Binding: 'Apparel',
+            Brand: 'Pearl iZUMi',
+            Department: 'mens',
+            Feature: [
+              'Select transfer fabric sets the benchmark for moisture transfer and four-way performance stretch',
+              '6-Panel anatomic design for superior, chafe-free comfort, UPF 50+ sun protection',
+              'Comfortable silicone leg grippers keep shorts in place, moisture-wicking, antimicrobial properties',
+              'Tour 3D Chamois is male-specific and non-chafing with padding in key areas',
+              '86 percent nylon, 14% spandex, 9-Inch inseam'
+            ],
+            ItemDimensions: {
+              Height: {
+                Units: 'inches',
+                Value: '2.00'
+              },
+              Length: {
+                Units: 'inches',
+                Value: '9.00'
+              },
+              Width: {
+                Units: 'inches',
+                Value: '9.00'
+              }
+            },
+            Label: 'Pearl iZUMi',
+            ListPrice: {
+              Amount: '50.00',
+              CurrencyCode: 'USD'
+            },
+            Manufacturer: 'Pearl iZUMi',
+            Model: '0275',
+            PackageDimensions: {
+              Height: {
+                Units: 'inches',
+                Value: '2.80'
+              },
+              Length: {
+                Units: 'inches',
+                Value: '9.75'
+              },
+              Width: {
+                Units: 'inches',
+                Value: '8.40'
+              },
+              Weight: {
+                Units: 'pounds',
+                Value: '0.40'
+              }
+            },
+            PackageQuantity: '1',
+            PartNumber: '0275',
+            ProductGroup: 'Apparel',
+            ProductTypeName: 'SHORTS',
+            Publisher: 'Pearl iZUMi',
+            SmallImage: {
+              URL: 'http://ecx.images-amazon.com/images/I/41ty3Sn%2BU8L._SL75_.jpg',
+              Height: {
+                Units: 'pixels',
+                Value: '75'
+              },
+              Width: {
+                Units: 'pixels',
+                Value: '58'
+              }
+            },
+            Studio: 'Pearl iZUMi',
+            Title: 'Pearl iZUMi Men\'s Quest Cycling Short'
+          }
+        },
+        Relationships: {
+          VariationChild: [
+            {
+              Identifiers: {
+                MarketplaceASIN: {
+                  MarketplaceId: 'ATVPDKIKX0DER',
+                  ASIN: 'B002KT3XQC'
+                }
+              },
+              Color: 'Black',
+              Size: 'Small'
+            },
+            {
+              Identifiers: {
+                MarketplaceASIN: {
+                  MarketplaceId: 'ATVPDKIKX0DER',
+                  ASIN: 'B002KT3XQW'
+                }
+              },
+              Color: 'Black',
+              Size: 'Medium'
+            },
+            {
+              Identifiers: {
+                MarketplaceASIN: {
+                  MarketplaceId: 'ATVPDKIKX0DER',
+                  ASIN: 'B002KT3XQM'
+                }
+              },
+              Color: 'Black',
+              Size: 'Large'
+            },
+            {
+              Identifiers: {
+                MarketplaceASIN: {
+                  MarketplaceId: 'ATVPDKIKX0DER',
+                  ASIN: 'B002KT3XR6'
+                }
+              },
+              Color: 'Black',
+              Size: 'X-Large'
+            },
+            {
+              Identifiers: {
+                MarketplaceASIN: {
+                  MarketplaceId: 'ATVPDKIKX0DER',
+                  ASIN: 'B002KT3XRG'
+                }
+              },
+              Color: 'Black',
+              Size: 'XX-Large'
+            }
+          ]
+        },
+        SalesRankings: {
+          SalesRank: [
+            {
+              ProductCategoryId: 'apparel_display_on_website',
+              Rank: '159'
+            },
+            {
+              ProductCategoryId: '2420095011',
+              Rank: '1'
+            },
+            {
+              ProductCategoryId: '2611189011',
+              Rank: '4'
+            }
+          ]
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: 'b12caada-d330-4d87-a789-EXAMPLE35872'
+    }
+  },
+  getMatchingProductForId: {
+    GetMatchingProductForIdResult: [
+      {
+        Id: '9781933988665',
+        IdType: 'ISBN',
+        status: 'Success',
+        Products: {
+          Product: {
+            Identifiers: {
+              MarketplaceASIN: {
+                MarketplaceId: 'ATVPDKIKX0DER',
+                ASIN: '1933988665'
+              }
+            },
+            AttributeSets: {
+              ItemAttributes: {
+                lang: 'en-US',
+                Author: [
+                  'Marmanis, Haralambos',
+                  'Babenko, Dmitry'
+                ],
+                Binding: 'Paperback',
+                Edition: '1',
+                ItemDimensions: {
+                  Height: {
+                    Units: 'inches',
+                    Value: '9.17'
+                  },
+                  Length: {
+                    Units: 'inches',
+                    Value: '7.36'
+                  },
+                  Width: {
+                    Units: 'inches',
+                    Value: '0.75'
+                  },
+                  Weight: {
+                    Units: 'pounds',
+                    Value: '1.40'
+                  }
+                },
+                IsEligibleForTradeIn: true,
+                Label: 'Manning Publications',
+                Languages: {
+                  Language: [
+                    {
+                      Name: 'english',
+                      Type: 'Unknown'
+                    },
+                    {
+                      Name: 'english',
+                      Type: 'Original Language'
+                    },
+                    {
+                      Name: 'english',
+                      Type: 'Published'
+                    }
+                  ]
+                },
+                ListPrice: {
+                  Amount: '44.99',
+                  CurrencyCode: 'USD'
+                },
+                Manufacturer: 'Manning Publications',
+                NumberOfItems: '1',
+                NumberOfPages: '368',
+                PackageDimensions: {
+                  Height: {
+                    Units: 'inches',
+                    Value: '0.80'
+                  },
+                  Length: {
+                    Units: 'inches',
+                    Value: '9.10'
+                  },
+                  Width: {
+                    Units: 'inches',
+                    Value: '7.30'
+                  },
+                  Weight: {
+                    Units: 'pounds',
+                    Value: '1.35'
+                  }
+                },
+                ProductGroup: 'Book',
+                ProductTypeName: 'ABIS_BOOK',
+                PublicationDate: '2009-07-05',
+                Publisher: 'Manning Publications',
+                SmallImage: {
+                  URL: 'http://ecx.images-amazon.com/images/I/51EEz05N2HL._SL75_.jpg',
+                  Height: {
+                    Units: 'pixels',
+                    Value: '75'
+                  },
+                  Width: {
+                    Units: 'pixels',
+                    Value: '60'
+                  }
+                },
+                Studio: 'Manning Publications',
+                Title: 'Algorithms of the Intelligent Web'
+              }
+            },
+            Relationships: null,
+            SalesRankings: {
+              SalesRank: [
+                {
+                  ProductCategoryId: 'book_display_on_website',
+                  Rank: '59485'
+                },
+                {
+                  ProductCategoryId: '377886011',
+                  Rank: '32'
+                },
+                {
+                  ProductCategoryId: '3887',
+                  Rank: '66'
+                },
+                {
+                  ProductCategoryId: '3870',
+                  Rank: '82'
+                }
+              ]
+            }
+          }
+        }
+      },
+      {
+        Id: '0439708184',
+        IdType: 'ISBN',
+        status: 'Success',
+        Products: {
+          Product: [
+            {
+              Identifiers: {
+                MarketplaceASIN: {
+                  MarketplaceId: 'ATVPDKIKX0DER',
+                  ASIN: '059035342X'
+                }
+              },
+              AttributeSets: {
+                ItemAttributes: {
+                  lang: 'en-US',
+                  Author: 'Rowling, J.K.',
+                  Binding: 'Paperback',
+                  Brand: 'Scholastic Press',
+                  Creator: {
+                    Role: 'Illustrator',
+                    Value: 'GrandPrÃ©, Mary'
+                  },
+                  Edition: '1st',
+                  Feature: 'Recommended Age: 9 years and up',
+                  ItemDimensions: {
+                    Height: {
+                      Units: 'inches',
+                      Value: '0.80'
+                    },
+                    Length: {
+                      Units: 'inches',
+                      Value: '7.50'
+                    },
+                    Width: {
+                      Units: 'inches',
+                      Value: '5.20'
+                    },
+                    Weight: {
+                      Units: 'pounds',
+                      Value: '0.50'
+                    }
+                  },
+                  IsAutographed: false,
+                  IsMemorabilia: false,
+                  Label: 'Scholastic Paperbacks',
+                  Languages: {
+                    Language: [
+                      {
+                        Name: 'english',
+                        Type: 'Unknown'
+                      },
+                      {
+                        Name: 'english',
+                        Type: 'Original Language'
+                      },
+                      {
+                        Name: 'english',
+                        Type: 'Published'
+                      }
+                    ]
+                  },
+                  ListPrice: {
+                    Amount: '10.99',
+                    CurrencyCode: 'USD'
+                  },
+                  Manufacturer: 'Scholastic Paperbacks',
+                  NumberOfItems: '1',
+                  NumberOfPages: '320',
+                  PackageDimensions: {
+                    Height: {
+                      Units: 'inches',
+                      Value: '1.00'
+                    },
+                    Length: {
+                      Units: 'inches',
+                      Value: '7.50'
+                    },
+                    Width: {
+                      Units: 'inches',
+                      Value: '5.20'
+                    },
+                    Weight: {
+                      Units: 'pounds',
+                      Value: '0.50'
+                    }
+                  },
+                  PackageQuantity: '1',
+                  PartNumber: '9780590353427',
+                  ProductGroup: 'Book',
+                  ProductTypeName: 'ABIS_BOOK',
+                  PublicationDate: '1999-10-01',
+                  Publisher: 'Scholastic Paperbacks',
+                  ReleaseDate: '1999-09-08',
+                  SmallImage: {
+                    URL: 'http://ecx.images-amazon.com/images/I/51MU5VilKpL._SL75_.jpg',
+                    Height: {
+                      Units: 'pixels',
+                      Value: '75'
+                    },
+                    Width: {
+                      Units: 'pixels',
+                      Value: '51'
+                    }
+                  },
+                  Studio: 'Scholastic Paperbacks',
+                  Title: 'Harry Potter and the Sorcerer\'s Stone (Book 1)'
+                }
+              },
+              Relationships: null,
+              SalesRankings: {
+                SalesRank: [
+                  {
+                    ProductCategoryId: 'book_display_on_website',
+                    Rank: '362'
+                  },
+                  {
+                    ProductCategoryId: '15356791',
+                    Rank: '6'
+                  },
+                  {
+                    ProductCategoryId: '3153',
+                    Rank: '9'
+                  },
+                  {
+                    ProductCategoryId: '3045',
+                    Rank: '14'
+                  }
+                ]
+              }
+            },
+            {
+              Identifiers: {
+                MarketplaceASIN: {
+                  MarketplaceId: 'ATVPDKIKX0DER',
+                  ASIN: '0439708184'
+                }
+              },
+              AttributeSets: {
+                ItemAttributes: {
+                  lang: 'en-US',
+                  Author: 'J.K. Rowling, Mary GrandPrÃƒÂ©',
+                  Binding: 'Paperback',
+                  Label: 'Scholastic Paperbacks',
+                  Languages: {
+                    Language: [
+                      {
+                        Name: 'english',
+                        Type: 'Unknown'
+                      },
+                      {
+                        Name: 'english',
+                        Type: 'Published'
+                      }
+                    ]
+                  },
+                  Manufacturer: 'Scholastic Paperbacks',
+                  NumberOfItems: '1',
+                  NumberOfPages: '784',
+                  PackageDimensions: {
+                    Height: {
+                      Units: 'inches',
+                      Value: '0.73'
+                    },
+                    Length: {
+                      Units: 'inches',
+                      Value: '8.43'
+                    },
+                    Width: {
+                      Units: 'inches',
+                      Value: '5.85'
+                    },
+                    Weight: {
+                      Units: 'pounds',
+                      Value: '0.79'
+                    }
+                  },
+                  ProductGroup: 'Book',
+                  ProductTypeName: 'BOOKS_1973_AND_LATER',
+                  PublicationDate: '1999-09-08',
+                  Publisher: 'Scholastic Paperbacks',
+                  SmallImage: {
+                    URL: 'http://ecx.images-amazon.com/images/I/51STfI7UiCL._SL75_.jpg',
+                    Height: {
+                      Units: 'pixels',
+                      Value: '75'
+                    },
+                    Width: {
+                      Units: 'pixels',
+                      Value: '52'
+                    }
+                  },
+                  Studio: 'Scholastic Paperbacks',
+                  Title: 'Harry Potter and the Sorcerer\'s Stone'
+                }
+              },
+              Relationships: null,
+              SalesRankings: {
+                SalesRank: {
+                  ProductCategoryId: 'book_display_on_website',
+                  Rank: '2091024'
+                }
+              }
+            }
+          ]
+        }
+      }
+    ],
+    ResponseMetadata: {
+      RequestId: '7ba3245e-a213-430a-bea9-EXAMPLE38d76'
+    }
+  },
+  getCompetitivePricingForSKU: {
+    GetCompetitivePricingForSKUResult: {
+      SellerSKU: 'SKU2468',
+      status: 'Success',
+      Product: {
+        Identifiers: {
+          MarketplaceASIN: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            ASIN: '1933890517'
+          },
+          SKUIdentifier: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            SellerId: 'A1IMEXAMPLEWRC',
+            SellerSKU: 'SKU2468'
+          }
+        },
+        CompetitivePricing: {
+          CompetitivePrices: {
+            CompetitivePrice: [
+              {
+                belongsToRequester: false,
+                condition: 'New',
+                subcondition: 'New',
+                CompetitivePriceId: '1',
+                Price: {
+                  LandedPrice: {
+                    CurrencyCode: 'USD',
+                    Amount: '40.03'
+                  },
+                  ListingPrice: {
+                    CurrencyCode: 'USD',
+                    Amount: '40.03'
+                  },
+                  Shipping: {
+                    CurrencyCode: 'USD',
+                    Amount: '0.00'
+                  }
+                }
+              },
+              {
+                belongsToRequester: false,
+                condition: 'Used',
+                subcondition: 'Good',
+                CompetitivePriceId: '2',
+                Price: {
+                  LandedPrice: {
+                    CurrencyCode: 'USD',
+                    Amount: '30.50'
+                  },
+                  ListingPrice: {
+                    CurrencyCode: 'USD',
+                    Amount: '30.50'
+                  },
+                  Shipping: {
+                    CurrencyCode: 'USD',
+                    Amount: '0.00'
+                  }
+                }
+              }
+            ]
+          },
+          NumberOfOfferListings: {
+            OfferListingCount: [
+              {
+                condition: 'Any',
+                Value: '296'
+              },
+              {
+                condition: 'Used',
+                Value: '145'
+              },
+              {
+                condition: 'New',
+                Value: '151'
+              }
+            ]
+          },
+          TradeInValue: {
+            CurrencyCode: 'USD',
+            Amount: '17.05'
+          }
+        },
+        SalesRankings: {
+          SalesRank: [
+            {
+              ProductCategoryId: 'book_display_on_website',
+              Rank: '900'
+            },
+            {
+              ProductCategoryId: '271578011',
+              Rank: '1'
+            },
+            {
+              ProductCategoryId: '355562011',
+              Rank: '8'
+            },
+            {
+              ProductCategoryId: '173516',
+              Rank: '25'
+            }
+          ]
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: '003fc73e-bf70-4e55-8db8-EXAMPLEce417'
+    }
+  },
+  getCompetitivePricingForASIN: {
+    GetCompetitivePricingForASINResult: {
+      ASIN: 'B002KT3XQM',
+      status: 'Success',
+      Product: {
+        Identifiers: {
+          MarketplaceASIN: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            ASIN: 'B002KT3XQM'
+          }
+        },
+        CompetitivePricing: {
+          CompetitivePrices: {
+            CompetitivePrice: {
+              belongsToRequester: false,
+              condition: 'New',
+              subcondition: 'New',
+              CompetitivePriceId: '1',
+              Price: {
+                LandedPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '34.27'
+                },
+                ListingPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '34.27'
+                },
+                Shipping: {
+                  CurrencyCode: 'USD',
+                  Amount: '0.00'
+                }
+              }
+            }
+          },
+          NumberOfOfferListings: {
+            OfferListingCount: [
+              {
+                condition: 'Any',
+                Value: '11'
+              },
+              {
+                condition: 'Used',
+                Value: '1'
+              },
+              {
+                condition: 'New',
+                Value: '10'
+              }
+            ]
+          }
+        },
+        SalesRankings: null
+      }
+    },
+    ResponseMetadata: {
+      RequestId: 'a8ec20c3-fc88-46e3-a128-EXAMPLE3fcdf'
+    }
+  },
+  getLowestOfferListingsForSKU: {
+    GetLowestOfferListingsForSKUResult: {
+      SellerSKU: 'SKU2468',
+      status: 'Success',
+      AllOfferListingsConsidered: false,
+      Product: {
+        Identifiers: {
+          MarketplaceASIN: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            ASIN: '1933890517'
+          },
+          SKUIdentifier: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            SellerId: 'A1IMEXAMPLEWRC',
+            SellerSKU: 'SKU2468'
+          }
+        },
+        LowestOfferListings: {
+          LowestOfferListing: [
+            {
+              Qualifiers: {
+                ItemCondition: 'Used',
+                ItemSubcondition: 'Acceptable',
+                FulfillmentChannel: 'Merchant',
+                ShipsDomestically: true,
+                ShippingTime: {
+                  Max: '0-2 days'
+                },
+                SellerPositiveFeedbackRating: '95-97%'
+              },
+              NumberOfOfferListingsConsidered: '3',
+              SellerFeedbackCount: '8900',
+              Price: {
+                LandedPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '28.68'
+                },
+                ListingPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '24.69'
+                },
+                Shipping: {
+                  CurrencyCode: 'USD',
+                  Amount: '3.99'
+                }
+              },
+              MultipleOffersAtLowestPrice: true
+            },
+            {
+              Qualifiers: {
+                ItemCondition: 'Used',
+                ItemSubcondition: 'Good',
+                FulfillmentChannel: 'Amazon',
+                ShipsDomestically: true,
+                ShippingTime: {
+                  Max: '0-2 days'
+                },
+                SellerPositiveFeedbackRating: '90-94%'
+              },
+              NumberOfOfferListingsConsidered: '1',
+              SellerFeedbackCount: '1569694',
+              Price: {
+                LandedPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '30.50'
+                },
+                ListingPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '30.50'
+                },
+                Shipping: {
+                  CurrencyCode: 'USD',
+                  Amount: '0.00'
+                }
+              },
+              MultipleOffersAtLowestPrice: false
+            },
+            {
+              Qualifiers: {
+                ItemCondition: 'Used',
+                ItemSubcondition: 'Good',
+                FulfillmentChannel: 'Merchant',
+                ShipsDomestically: true,
+                ShippingTime: {
+                  Max: '0-2 days'
+                },
+                SellerPositiveFeedbackRating: '95-97%'
+              },
+              NumberOfOfferListingsConsidered: '3',
+              SellerFeedbackCount: '7732',
+              Price: {
+                LandedPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '30.99'
+                },
+                ListingPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '27.00'
+                },
+                Shipping: {
+                  CurrencyCode: 'USD',
+                  Amount: '3.99'
+                }
+              },
+              MultipleOffersAtLowestPrice: false
+            }
+          ]
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: '8106ed19-9a90-4820-881e-EXAMPLE0ab97'
+    }
+  },
+  getLowestOfferListingsForASIN: {
+    GetLowestOfferListingsForASINResult: {
+      ASIN: 'B002KT3XQM',
+      status: 'Success',
+      AllOfferListingsConsidered: true,
+      Product: {
+        Identifiers: {
+          MarketplaceASIN: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            ASIN: 'B002KT3XQM'
+          }
+        },
+        LowestOfferListings: {
+          LowestOfferListing: [
+            {
+              Qualifiers: {
+                ItemCondition: 'Used',
+                ItemSubcondition: 'VeryGood',
+                FulfillmentChannel: 'Merchant',
+                ShipsDomestically: true,
+                ShippingTime: {
+                  Max: '0-2 days'
+                },
+                SellerPositiveFeedbackRating: '90-94%'
+              },
+              NumberOfOfferListingsConsidered: '1',
+              SellerFeedbackCount: '762',
+              Price: {
+                LandedPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '32.99'
+                },
+                ListingPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '28.00'
+                },
+                Shipping: {
+                  CurrencyCode: 'USD',
+                  Amount: '4.99'
+                }
+              },
+              MultipleOffersAtLowestPrice: false
+            },
+            {
+              Qualifiers: {
+                ItemCondition: 'New',
+                ItemSubcondition: 'New',
+                FulfillmentChannel: 'Amazon',
+                ShipsDomestically: true,
+                ShippingTime: {
+                  Max: '0-2 days'
+                },
+                SellerPositiveFeedbackRating: '98-100%'
+              },
+              NumberOfOfferListingsConsidered: '1',
+              SellerFeedbackCount: '181744',
+              Price: {
+                LandedPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '34.27'
+                },
+                ListingPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '34.27'
+                },
+                Shipping: {
+                  CurrencyCode: 'USD',
+                  Amount: '0.00'
+                }
+              },
+              MultipleOffersAtLowestPrice: false
+            },
+            {
+              Qualifiers: {
+                ItemCondition: 'New',
+                ItemSubcondition: 'New',
+                FulfillmentChannel: 'Amazon',
+                ShipsDomestically: true,
+                ShippingTime: {
+                  Max: '0-2 days'
+                },
+                SellerPositiveFeedbackRating: '95-97%'
+              },
+              NumberOfOfferListingsConsidered: '1',
+              SellerFeedbackCount: '13213',
+              Price: {
+                LandedPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '41.18'
+                },
+                ListingPrice: {
+                  CurrencyCode: 'USD',
+                  Amount: '41.18'
+                },
+                Shipping: {
+                  CurrencyCode: 'USD',
+                  Amount: '0.00'
+                }
+              },
+              MultipleOffersAtLowestPrice: false
+            }
+          ]
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: '60979901-82af-457b-8bdd-EXAMPLE28478'
+    }
+  },
+  getLowestPricedOffersForSKU: {
+    GetLowestPricedOffersForSKUResult: {
+      MarketplaceID: 'ATVPDKIKX0DER',
+      SKU: 'GE Product',
+      ItemCondition: 'New',
+      status: 'Success',
+      Identifier: {
+        MarketplaceId: 'ATVPDKIKX0DER',
+        SellerSKU: 'GE Product',
+        ItemCondition: 'New',
+        TimeOfOfferChange: '2015-07-19T23:15:11.859Z'
+      },
+      Summary: {
+        TotalOfferCount: '1',
+        NumberOfOffers: {
+          OfferCount: {
+            condition: 'new',
+            fulfillmentChannel: 'Amazon',
+            Value: '1'
+          }
+        },
+        LowestPrices: {
+          LowestPrice: {
+            condition: 'new',
+            fulfillmentChannel: 'Amazon',
+            LandedPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            ListingPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            Shipping: {
+              CurrencyCode: 'USD',
+              Amount: '0.00'
+            }
+          }
+        },
+        BuyBoxPrices: {
+          BuyBoxPrice: {
+            condition: 'New',
+            LandedPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            ListingPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            Shipping: {
+              CurrencyCode: 'USD',
+              Amount: '0.00'
+            }
+          }
+        },
+        ListPrice: {
+          CurrencyCode: 'USD',
+          Amount: '58.34'
+        },
+        SuggestedLowerPricePlusShipping: {
+          CurrencyCode: 'USD',
+          Amount: '32.99'
+        },
+        BuyBoxEligibleOffers: {
+          OfferCount: {
+            condition: 'new',
+            fulfillmentChannel: 'Amazon',
+            Value: '1'
+          }
+        }
+      },
+      Offers: {
+        Offer: {
+          MyOffer: false,
+          SubCondition: 'new',
+          SellerFeedbackRating: {
+            SellerPositiveFeedbackRating: '100.0',
+            FeedbackCount: '1'
+          },
+          ShippingTime: {
+            minimumHours: '0',
+            maximumHours: '0',
+            availabilityType: 'NOW'
+          },
+          ListingPrice: {
+            CurrencyCode: 'USD',
+            Amount: '32.99'
+          },
+          Shipping: {
+            CurrencyCode: 'USD',
+            Amount: '0.00'
+          },
+          IsFulfilledByAmazon: true,
+          IsBuyBoxWinner: true,
+          IsFeaturedMerchant: true
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: 'fd2e6c6d-0b6d-499b-9fea-074175c0547a'
+    }
+  },
+  getLowestPricedOffersForASIN: {
+    GetLowestPricedOffersForASINResult: {
+      MarketplaceID: 'ATVPDKIKX0DER',
+      ItemCondition: 'New',
+      ASIN: 'B0002GTTRC',
+      status: 'Success',
+      Identifier: {
+        MarketplaceId: 'ATVPDKIKX0DER',
+        ASIN: 'B0002GTTRC',
+        ItemCondition: 'New',
+        TimeOfOfferChange: '2015-07-19T23:15:11.859Z'
+      },
+      Summary: {
+        TotalOfferCount: '0',
+        NumberOfOffers: {
+          OfferCount: {
+            condition: 'new',
+            fulfillmentChannel: 'Amazon',
+            Value: '1'
+          }
+        },
+        LowestPrices: {
+          LowestPrice: {
+            condition: 'new',
+            fulfillmentChannel: 'Amazon',
+            LandedPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            ListingPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            Shipping: {
+              CurrencyCode: 'USD',
+              Amount: '0.00'
+            }
+          }
+        },
+        BuyBoxPrices: {
+          BuyBoxPrice: {
+            condition: 'New',
+            LandedPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            ListingPrice: {
+              CurrencyCode: 'USD',
+              Amount: '32.99'
+            },
+            Shipping: {
+              CurrencyCode: 'USD',
+              Amount: '0.00'
+            }
+          }
+        },
+        ListPrice: {
+          CurrencyCode: 'USD',
+          Amount: '58.34'
+        },
+        SuggestedLowerPricePlusShipping: {
+          CurrencyCode: 'USD',
+          Amount: '32.99'
+        },
+        BuyBoxEligibleOffers: {
+          OfferCount: {
+            condition: 'new',
+            fulfillmentChannel: 'Amazon',
+            Value: '1'
+          }
+        }
+      },
+      Offers: {
+        Offer: {
+          SubCondition: 'new',
+          SellerFeedbackRating: {
+            SellerPositiveFeedbackRating: '100.0',
+            FeedbackCount: '1'
+          },
+          ShippingTime: {
+            minimumHours: '0',
+            maximumHours: '0',
+            availabilityType: 'NOW'
+          },
+          ListingPrice: {
+            CurrencyCode: 'USD',
+            Amount: '32.99'
+          },
+          Shipping: {
+            CurrencyCode: 'USD',
+            Amount: '0.00'
+          },
+          IsFulfilledByAmazon: true,
+          IsBuyBoxWinner: true,
+          IsFeaturedMerchant: true
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: '21534f03-e7e3-4d9a-9bea-0e15add3326d'
+    }
+  },
+  getMyFeesEstimate: {
+    GetMyFeesEstimateResult: {
+      FeesEstimateResultList: {
+        FeesEstimateResult: {
+          Status: 'Success',
+          FeesEstimateIdentifier: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            IdValue: 'B0002GTTRC',
+            IdType: 'ASIN',
+            IsAmazonFulfilled: true,
+            PriceToEstimateFees: {
+              ListingPrice: {
+                CurrencyCode: 'USD',
+                Amount: '58.00'
+              },
+              Shipping: {
+                CurrencyCode: 'USD',
+                Amount: '0.00'
+              },
+              Points: {
+                PointsNumber: '0'
+              }
+            },
+            SellerInputIdentifier: 'IDDDDDDDD'
+          },
+          FeesEstimate: {
+            TimeOfFeesEstimation: '2015-07-19T23:15:11.859Z',
+            TotalFeesEstimate: {
+              CurrencyCode: 'USD',
+              Amount: '10.00'
+            },
+            FeeDetailList: {
+              FeeDetail: [
+                {
+                  FeeType: 'AmazonReferralFee',
+                  FeeAmount: {
+                    CurrencyCode: 'USD',
+                    Amount: '8.70'
+                  },
+                  FeePromotion: {
+                    CurrencyCode: 'USD',
+                    Amount: '1.00'
+                  },
+                  FinalFee: {
+                    CurrencyCode: 'USD',
+                    Amount: '7.70'
+                  }
+                },
+                {
+                  FeeType: 'VariableClosingFee',
+                  FeeAmount: {
+                    CurrencyCode: 'USD',
+                    Amount: '0.00'
+                  },
+                  FinalFee: {
+                    CurrencyCode: 'USD',
+                    Amount: '0.0'
+                  }
+                },
+                {
+                  FeeType: 'FulfillmentFees',
+                  FeeAmount: {
+                    CurrencyCode: 'USD',
+                    Amount: '2.30'
+                  },
+                  FinalFee: {
+                    CurrencyCode: 'USD',
+                    Amount: '2.30'
+                  },
+                  IncludedFeeList: {
+                    FeeDetail: [
+                      {
+                        FeeType: 'OrderHandlingFee',
+                        FeeAmount: {
+                          CurrencyCode: 'USD',
+                          Amount: '1.00'
+                        },
+                        FinalFee: {
+                          CurrencyCode: 'USD',
+                          Amount: '1.00'
+                        }
+                      },
+                      {
+                        FeeType: 'PickAndPackFee',
+                        FeeAmount: {
+                          CurrencyCode: 'USD',
+                          Amount: '0.30'
+                        },
+                        FinalFee: {
+                          CurrencyCode: 'USD',
+                          Amount: '0.30'
+                        }
+                      },
+                      {
+                        FeeType: 'WeightHandlingFee',
+                        FeeAmount: {
+                          CurrencyCode: 'USD',
+                          Amount: '1.00'
+                        },
+                        FinalFee: {
+                          CurrencyCode: 'USD',
+                          Amount: '1.00'
+                        }
+                      }
+                    ]
+                  }
+                }
+              ]
+            }
+          }
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: '21534f03-e7e3-4d9a-9bea-0e15add3326d'
+    }
+  },
+  getMyPriceForSKU: {
+    GetMyPriceForSKUResult: {
+      SellerSKU: 'SKU2468',
+      status: 'Success',
+      Product: {
+        Identifiers: {
+          MarketplaceASIN: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            ASIN: '1933890517'
+          },
+          SKUIdentifier: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            SellerId: 'A1IMEXAMPLEWRC',
+            SellerSKU: 'SKU2468'
+          }
+        },
+        Offers: {
+          Offer: {
+            BuyingPrice: {
+              LandedPrice: {
+                CurrencyCode: 'USD',
+                Amount: '303.99'
+              },
+              ListingPrice: {
+                CurrencyCode: 'USD',
+                Amount: '300.00'
+              },
+              Shipping: {
+                CurrencyCode: 'USD',
+                Amount: '3.99'
+              }
+            },
+            RegularPrice: {
+              CurrencyCode: 'USD',
+              Amount: '300.00'
+            },
+            FulfillmentChannel: 'MERCHANT',
+            ItemCondition: 'Used',
+            ItemSubCondition: 'Acceptable',
+            SellerId: 'A1IMEXAMPLEWRC',
+            SellerSKU: 'SKU2468'
+          }
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: 'bc6e4601-3d74-4612-adcf-EXAMPLEf1796'
+    }
+  },
+  getMyPriceForASIN: {
+    GetMyPriceForASINResult: {
+      ASIN: '1933890517',
+      status: 'Success',
+      Product: {
+        Identifiers: {
+          MarketplaceASIN: {
+            MarketplaceId: 'ATVPDKIKX0DER',
+            ASIN: '1933890517'
+          }
+        },
+        Offers: {
+          Offer: {
+            BuyingPrice: {
+              LandedPrice: {
+                CurrencyCode: 'USD',
+                Amount: '303.99'
+              },
+              ListingPrice: {
+                CurrencyCode: 'USD',
+                Amount: '300.00'
+              },
+              Shipping: {
+                CurrencyCode: 'USD',
+                Amount: '3.99'
+              }
+            },
+            RegularPrice: {
+              CurrencyCode: 'USD',
+              Amount: '300.00'
+            },
+            FulfillmentChannel: 'MERCHANT',
+            ItemCondition: 'Used',
+            ItemSubCondition: 'Acceptable',
+            SellerId: 'A1IMEXAMPLEWRC',
+            SellerSKU: 'SKU2468'
+          }
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: 'a3381684-87bd-416e-9b95-EXAMPLE9c236'
+    }
+  },
+  getProductCategoriesForSKU: {
+    GetProductCategoriesForSKUResult: {
+      Self: [
+        {
+          ProductCategoryId: '271578011',
+          ProductCategoryName: 'Project Management',
+          Parent: {
+            ProductCategoryId: '2675',
+            ProductCategoryName: 'Management & Leadership',
+            Parent: {
+              ProductCategoryId: '3',
+              ProductCategoryName: 'Business & Investing',
+              Parent: {
+                ProductCategoryId: '1000',
+                ProductCategoryName: 'Subjects',
+                Parent: {
+                  ProductCategoryId: '283155',
+                  ProductCategoryName: 'Subjects'
+                }
+              }
+            }
+          }
+        },
+        {
+          ProductCategoryId: '684248011',
+          ProductCategoryName: 'Management',
+          Parent: {
+            ProductCategoryId: '468220',
+            ProductCategoryName: 'Business & Finance',
+            Parent: {
+              ProductCategoryId: '465600',
+              ProductCategoryName: 'New, Used & Rental Textbooks',
+              Parent: {
+                ProductCategoryId: '2349030011',
+                ProductCategoryName: 'Specialty Boutique',
+                Parent: {
+                  ProductCategoryId: '283155',
+                  ProductCategoryName: 'Specialty Boutique'
+                }
+              }
+            }
+          }
+        }
+      ]
+    },
+    ResponseMetadata: {
+      RequestId: 'e058aabd-b4c3-48ba-9bfa-EXAMPLE9a267'
+    }
+  },
+  getProductCategoriesForASIN: {
+    GetProductCategoriesForASINResult: {
+      Self: {
+        ProductCategoryId: '2420095011',
+        ProductCategoryName: 'Compression Shorts',
+        Parent: {
+          ProductCategoryId: '2419332011',
+          ProductCategoryName: 'Men',
+          Parent: {
+            ProductCategoryId: '2371051011',
+            ProductCategoryName: 'Clothing',
+            Parent: {
+              ProductCategoryId: '3403201',
+              ProductCategoryName: 'Bikes & Accessories',
+              Parent: {
+                ProductCategoryId: '2232464011',
+                ProductCategoryName: 'Bikes & Scooters',
+                Parent: {
+                  ProductCategoryId: '3375301',
+                  ProductCategoryName: 'Categories',
+                  Parent: {
+                    ProductCategoryId: '3375251',
+                    ProductCategoryName: 'Categories'
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: 'fbce5b62-67cc-4ab8-86f3-EXAMPLE22e4e'
+    }
+  },
+  getServiceStatus: {
+    GetServiceStatusResponse: {
+      xmlns: 'http://mws.amazonservices.com/schema/Products/2011-10-01',
+      GetServiceStatusResult: { Status: 'GREEN', Timestamp: '2020-05-15T15:24:52.256Z' },
+      ResponseMetadata: { RequestId: 'ce35c00f-274f-487a-8180-2ce0fa8fb36a' }
+    }
+  }
+}
