+++ api/Sellers/mockup-data.js
@@ -1,0 +1,39 @@
+module.exports = {
+  listMarketplaceParticipations: {
+    ListMarketplaceParticipationsResult: {
+      NextToken: 'MRgZW55IGNhcm5hbCBwbGVhc3VyZS6=',
+      ListParticipations: {
+        Participation: {
+          MarketplaceId: 'ATVPDKIKX0DER',
+          SellerId: 'A135KKEKJAIBJ56',
+          HasSellerSuspendedListings: 'No'
+        }
+      },
+      ListMarketplaces: {
+        Marketplace: {
+          MarketplaceId: 'ATVPDKIKX0DER',
+          Name: 'Amazon.com',
+          DefaultCountryCode: 'US',
+          DefaultCurrencyCode: 'USD',
+          DefaultLanguageCode: 'en_US',
+          DomainName: 'www.amazon.com'
+        }
+      }
+    },
+    ResponseMetadata: {
+      RequestId: 'efeab958-74e2-45d4-9018-2323084413b5'
+    }
+  },
+  getServiceStatus: {
+    GetServiceStatusResponse: {
+      xmlns: 'https://mws.amazonservices.com/Sellers/2011-07-01',
+      GetServiceStatusResult: {
+        Status: 'GREEN',
+        Timestamp: '2020-05-15T15:51:29.295Z'
+      },
+      ResponseMetadata: {
+        RequestId: 'd8d121a3-ead0-4093-929e-dda74ef9e1e0'
+      }
+    }
+  }
+}
